requirements.txt
apache-airflow
pyarrow
psycopg2-binary
requests
warcio
pandas
